refactor: library.reader 이동에 따른 관련 테스트 패키지 이동
